requirements.txt
google-generativeai
pymupdf
markdown
tqdm
python-dotenv
markdown
weasyprint